+++ NEWS.md
@@ -1,4 +1,5 @@
 # News
 
 * 20250226 - CRAN submission
+* 20250302 - Added better space feature for x axis ticks to match the chromosome length of the dataset. Added several more specific parameters like two critical thresholds.